// front/src/typechain/MusicShop.ts
export declare namespace MusicShop {
  export type AlbumStruct = {
    index: PromiseOrValue<BigNumberish>;
    uuid: PromiseOrValue<string>;
    title: PromiseOrValue<string>;
    price: PromiseOrValue<BigNumberish>;
    quantity: PromiseOrValue<BigNumberish>;
  };

  export type AlbumStructOutput = [
    BigNumber,
    string,
    string,
    BigNumber,
    BigNumber
  ] & {
    index: BigNumber;
    uuid: string;
    title: string;
    price: BigNumber;
    quantity: BigNumber;
  };
}